Auto stash before merge of "master" and "origin/master"

+++ assn1/TVector_speical.h
@@ -10,7 +10,7 @@ public:
 	T2Double(double a, double b) { (*this)[0] = a, (*this)[1] = b; }
 	T2Double(const T2Double& x) { TVector<double, 2>::operator=(x); }
 	T2Double(const TVector<double, 2>& x) { TVector<double, 2>::operator=(x); }
-	T2Double(const Vec2f& x) { (*this)[0] = x[0], (*this)[1] = x[1]; }
+	T2Double(const Vec2d& x) { (*this)[0] = x[0], (*this)[1] = x[1]; }
 	void operator=(const T2Double& x) { TVector<double, 2>::operator=(x); }
-	Vec2f convert_gl(void) { return Vec2f((*this)[0], (*this)[1]); }
+	Vec2d convert_gl(void) { return Vec2d((*this)[0], (*this)[1]); }
 };
